safari-600.7-branch/LayoutTests/svg/animations/remove-animation-element-while-animation-is-running-expected.txt

SVG 1.1 dynamic animation tests


This removes an animation element while the animation is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect1.x.animVal.value is 50
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 50
PASS rect2.x.baseVal.value is 0
PASS rect1.x.animVal.value is 100
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 100
PASS rect2.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 100
PASS rect2.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 0
PASS rect1.x.animVal.value is 0
PASS rect1.x.baseVal.value is 0
PASS rect2.x.animVal.value is 0
PASS rect2.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
